# Conflicts:
#	code/game/machinery/vending.dm
#	code/game/objects/structures/crates_lockers/closets/secure/engineering.dm
#	code/game/objects/structures/crates_lockers/closets/secure/medical.dm
#	icons/mob/head.dmi
#	icons/obj/clothing/hats.dmi

+++ code/modules/food_and_drinks/food/snacks_meat.dm
@@ -95,10 +95,10 @@
 	tastes = list("meat" = 1, "salmon" = 1)
 	foodtype = MEAT | ALCOHOL
 
-/obj/item/reagent_containers/food/snacks/faggot
-	name = "faggot"
+/obj/item/reagent_containers/food/snacks/meatball
+	name = "meatball"
 	desc = "A great meal all round. Not a cord of wood."
-	icon_state = "faggot"
+	icon_state = "meatball"
 	list_reagents = list("nutriment" = 4, "vitamin" = 1)
 	filling_color = "#800000"
 	tastes = list("meat" = 1)
@@ -148,8 +148,8 @@
 	foodtype = VEGETABLES
 
 /obj/item/reagent_containers/food/snacks/kebab/tail
-	name = "lizard-tail kebab"
-	desc = "Severed lizard tail on a stick."
+	name = "unathi-tail kebab"
+	desc = "Severed unathi tail on a stick."
 	bonus_reagents = list("nutriment" = 1, "vitamin" = 4)
 	tastes = list("meat" = 8, "metal" = 4, "scales" = 1)
 	foodtype = MEAT
@@ -184,8 +184,14 @@
 	foodtype = MEAT | SUGAR
 
 /obj/item/reagent_containers/food/snacks/monkeycube/proc/Expand()
+	if(GLOB.total_cube_monkeys >= CONFIG_GET(number/max_cube_monkeys))
+		visible_message("<span class='warning'>[src] refuses to expand!</span>")
+		return
+
 	visible_message("<span class='notice'>[src] expands!</span>")
-	new /mob/living/carbon/monkey(get_turf(src))
+	var/mob/spammer = get_mob_by_key(src.fingerprintslast)
+	var/mob/living/carbon/monkey/cube/bananas = new(get_turf(src))
+	bananas.log_message("Spawned via [src] at [COORD(src)], Last attached mob: [key_name(spammer)].", INDIVIDUAL_ATTACK_LOG)
 	qdel(src)
 
 /obj/item/reagent_containers/food/snacks/enchiladas
